requirements.txt
langchain_core==0.3.7
langchain_groq==0.2.0
langchain_openai==0.2.1
pdfplumber==0.11.4
pydantic==2.9.2
python_pptx==1.0.2
streamlit==1.39.0
